Resolve issues with pictures not looking like previews. Try and fix some of the jumping around when going from taking a picture to viewing it. Catch places where the camera was causing crashes and address them. Stop stretching Camera Previews.
