+++ mobile/build.gradle
@@ -8,8 +8,8 @@ android {
         applicationId "ie.yesequality.yesequality"
         minSdkVersion 15
         targetSdkVersion 22
-        versionCode 2
-        versionName "1.0.0"
+        versionCode 3
+        versionName "1.0.1"
     }
 
     Properties props = new Properties()
